リーグ画面実装と既存リーグ画面をシーズンに移動
